Put back the vst3 casts that only look redundant on Windows
The `vst3` crate's constants are bindgen output, and bindgen takes the
integer type from the platform's C++ ABI: i32 on MSVC, u32 everywhere
else. Clippy on Windows calls `MediaTypes_::kAudio as i32` an
unnecessary cast and it is right — on Windows. Removing it broke the
Linux and macOS builds outright, twenty errors deep in vst3-host.

The casts are back and the lint is off for that crate, with the reason
written down next to it.

host-window's test modules are gated on `windows` as a whole now. Every
test in them already was, so on the other platforms their imports were
left looking unused.

+++ crates/vst3-host/src/vst_events.rs
@@ -142,12 +142,12 @@ fn to_vst_event(event: &NoteEvent) -> Option<VstEvent> {
 /// to control — so it is dropped until the mapping is read properly.
 fn expression_type_id(expression: NoteExpression) -> Option<u32> {
     Some(match expression {
-        NoteExpression::Volume => NoteExpressionTypeIDs_::kVolumeTypeID,
-        NoteExpression::Pan => NoteExpressionTypeIDs_::kPanTypeID,
-        NoteExpression::Tuning => NoteExpressionTypeIDs_::kTuningTypeID,
-        NoteExpression::Vibrato => NoteExpressionTypeIDs_::kVibratoTypeID,
-        NoteExpression::Expression => NoteExpressionTypeIDs_::kExpressionTypeID,
-        NoteExpression::Brightness => NoteExpressionTypeIDs_::kBrightnessTypeID,
+        NoteExpression::Volume => NoteExpressionTypeIDs_::kVolumeTypeID as u32,
+        NoteExpression::Pan => NoteExpressionTypeIDs_::kPanTypeID as u32,
+        NoteExpression::Tuning => NoteExpressionTypeIDs_::kTuningTypeID as u32,
+        NoteExpression::Vibrato => NoteExpressionTypeIDs_::kVibratoTypeID as u32,
+        NoteExpression::Expression => NoteExpressionTypeIDs_::kExpressionTypeID as u32,
+        NoteExpression::Brightness => NoteExpressionTypeIDs_::kBrightnessTypeID as u32,
         NoteExpression::Pressure => return None,
     })
 }
